Update Actions SDK documentation and amend other adapters to ensure shared lib can be used where possible.
