+++ libraries/Bot.Builder.Community.Adapters.ActionsSDK/Bot.Builder.Community.Adapters.ActionsSDK.csproj
@@ -5,6 +5,15 @@
     <TargetFramework>netstandard2.0</TargetFramework>
   </PropertyGroup>
 
+	<ItemGroup>
+	  <Compile Remove="action\**" />
+	  <Compile Remove="media\**" />
+	  <EmbeddedResource Remove="action\**" />
+	  <EmbeddedResource Remove="media\**" />
+	  <None Remove="action\**" />
+	  <None Remove="media\**" />
+	</ItemGroup>
+
   <ItemGroup>
     <PackageReference Include="Microsoft.Bot.Builder.Integration.AspNet.Core" Version="$(Bot_Builder_Version)" />
   </ItemGroup>
