+++ libraries/Bot.Builder.Community.Adapters.ActionsSDK/action/actions/actions.yaml
@@ -1,0 +1,2 @@
+custom:
+  actions.intent.MAIN: {}
