+++ libraries/Bot.Builder.Community.Adapters.ActionsSDK/action/custom/global/actions.intent.CANCEL.yaml
@@ -1,0 +1,3 @@
+handler:
+  webhookHandler: cancel
+transitionToScene: actions.scene.END_CONVERSATION
