+++ libraries/Bot.Builder.Community.Adapters.ActionsSDK/action/custom/global/actions.intent.MAIN.yaml
@@ -1,0 +1,3 @@
+handler:
+  webhookHandler: main
+transitionToScene: Prompt
