+++ libraries/Bot.Builder.Community.Adapters.ActionsSDK/action/custom/global/actions.intent.NO_INPUT_2.yaml
@@ -1,0 +1,2 @@
+handler:
+  webhookHandler: no_input2
