+++ libraries/Bot.Builder.Community.Adapters.ActionsSDK/action/custom/global/actions.intent.NO_INPUT_FINAL.yaml
@@ -1,0 +1,3 @@
+handler:
+  webhookHandler: no_input3
+transitionToScene: actions.scene.END_CONVERSATION
