+++ libraries/Bot.Builder.Community.Adapters.ActionsSDK/action/custom/global/actions.intent.NO_MATCH_1.yaml
@@ -1,0 +1,2 @@
+handler:
+  webhookHandler: no_match
