+++ libraries/Bot.Builder.Community.Adapters.ActionsSDK/action/custom/global/actions.intent.NO_MATCH_FINAL.yaml
@@ -1,0 +1,3 @@
+handler:
+  webhookHandler: no_match3
+transitionToScene: actions.scene.END_CONVERSATION
