+++ libraries/Bot.Builder.Community.Adapters.ActionsSDK/action/custom/intents/actions.intent.MAIN.yaml
@@ -1,0 +1,1 @@
+{}
