+++ libraries/Bot.Builder.Community.Adapters.ActionsSDK/action/custom/types/prompt_option.yaml
@@ -1,0 +1,2 @@
+synonym:
+  matchType: EXACT_MATCH
