+++ libraries/Bot.Builder.Community.Adapters.ActionsSDK/action/manifest.yaml
@@ -1,0 +1,1 @@
+version: "1.0"
