+++ libraries/Bot.Builder.Community.Adapters.Shared/MappingHelper.cs
@@ -20,6 +20,7 @@ namespace Bot.Builder.Community.Adapters.Shared
             }
 
             var activity = messageActivities.Last();
+            var endWithPeriod = activity.Text?.TrimEnd().EndsWith(".") ?? false;
 
             if (messageActivities.Any(a => !String.IsNullOrEmpty(a.Speak)))
             {
@@ -31,10 +32,15 @@ namespace Bot.Builder.Community.Adapters.Shared
                 activity.Speak = $"<speak>{speakText}</speak>";
             }
 
-            activity.Text = String.Join(". ", messageActivities
+            activity.Text = String.Join(" ", messageActivities
                 .Select(a => NormalizeActivityText(a.TextFormat, a.Text))
                 .Where(s => !String.IsNullOrEmpty(s))
-                .Select(s => s.Trim(new char[] { ' ', '.' })));
+                .Select(s => s.TrimEnd(' ')));
+
+            if (activity.Text.EndsWith(".") && !endWithPeriod)
+            {
+                activity.Text = activity.Text.TrimEnd('.');
+            }
 
             activity.Attachments = messageActivities.Where(x => x.Attachments != null).SelectMany(x => x.Attachments).ToList();
