+++ libraries/Bot.Builder.Community.Adapters.Zoom/Bot.Builder.Community.Adapters.Zoom.csproj
@@ -21,9 +21,19 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <Compile Remove="media\**" />
+    <EmbeddedResource Remove="media\**" />
+    <None Remove="media\**" />
+  </ItemGroup>
+
+  <ItemGroup>
     <PackageReference Include="Microsoft.Bot.Builder.Integration.AspNet.Core" Version="$(Bot_Builder_Version)" />
     <PackageReference Include="Microsoft.Net.Http.Headers" Version="2.1.1" />
     <PackageReference Include="RestSharp" Version="106.10.2-alpha.0.8" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Bot.Builder.Community.Adapters.Shared\Bot.Builder.Community.Adapters.Shared.csproj" />
   </ItemGroup>
 
 </Project>
