+++ libraries/Bot.Builder.Community.Adapters.Zoom/ZoomRequestMapper.cs
@@ -5,7 +5,6 @@ using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Logging.Abstractions;
 using System.Linq;
 using Bot.Builder.Community.Adapters.Zoom.Attachments;
-using Bot.Builder.Community.Adapters.Zoom.Helpers;
 
 namespace Bot.Builder.Community.Adapters.Zoom
 {
